src/main/java/com/github/ducknowledges/leetcodejavasolutions/easy/RemoveDuplicatesSortedArray.java
package com.github.ducknowledges.leetcodejavasolutions.easy;

/**
 * 26. Remove Duplicates from Sorted Array
 * https://leetcode.com/problems/remove-duplicates-from-sorted-array/
 * */
public class RemoveDuplicatesSortedArray {

    /*
     * Time: O(n)
     * Space: O(1)
     * */
    public int removeDuplicates(int[] nums) {
        int insertIndex = 1;
        for(int i = 1; i < nums.length; i++) {
            if (nums[i - 1] != nums[i]) {
                nums[insertIndex] = nums[i];
                insertIndex++;
            }
        }
        return insertIndex;
    }
}
